pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "exifsniffer"
version = "0.1.1"
description = "MCP Streamable HTTP server: fetch remote media, extract EXIF and container metadata as JSON lists"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "mcp>=1.27.0",
  "pydantic>=2.7",
  "httpx[http2]>=0.27.0",
  "pillow>=10.4.0",
  "piexif>=1.1.3",
  "uvicorn[standard]>=0.30.0",
  "starlette>=0.40.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
exifsniffer = "exifsniffer.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
